+++ dsmrreader/config/base.py
@@ -238,7 +238,6 @@ DSMRREADER_SUPPORTED_DB_VENDORS = ('postgresql', 'mysql')
 DSMRREADER_BACKUP_PG_DUMP = 'pg_dump'
 DSMRREADER_BACKUP_MYSQLDUMP = 'mysqldump'
 DSMRREADER_BACKUP_SQLITE = 'sqlite3'
-DSMRREADER_BACKUP_DIRECTORY = 'backups'  # Relative to project root.
 DSMRREADER_DROPBOX_SYNC_INTERVAL = 1  # Only check for changes once per hour.
 DSMRREADER_DROPBOX_ERROR_INTERVAL = 12  # Skip new files for 12 hours when insufficient space in Dropbox account.