+++ docs/changelog.rst
@@ -26,6 +26,7 @@ v1.27.0 - 2018-xx-xx
 - [`#560 <https://github.com/dennissiemensma/dsmr-reader/issues/560>`_] Added boundaryGap to improve charts - by jbrunink / Tijs van Noije
 - [`#561 <https://github.com/dennissiemensma/dsmr-reader/issues/561>`_] Arrows on status page will now be hidden on small screens where they don't make sense anymore - by jbrunink
 - [`#426 <https://github.com/dennissiemensma/dsmr-reader/issues/426>`_] Temperatuurmetingen per uur inzichtelijk als CSV
+- [`#558 <https://github.com/dennissiemensma/dsmr-reader/issues/558>`_] Custom backup storage location
 - [`#xxxxxxxx <https://github.com/dennissiemensma/dsmr-reader/issues/xxxxxxxx>`_] yyyyyyyyyyyyy